backend/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
apispec==4.6.0
APScheduler==3.9.1
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
bidict==0.21.4
blinker==1.4
cachelib==0.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
drf-yasg==1.20.0
eventlet==0.33.0
filelock==3.0.12
Flask==2.0.1
flask-apispec==0.11.0
Flask-APScheduler==1.12.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.12.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
packaging==21.0
Pillow==8.3.2
protobuf==3.19.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.15.0
PyJWT==1.7.1
pymongo==3.12.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
quick-mailer==2022.2.22
rdflib==5.0.0
redis==4.0.2
requests==2.25.1
requests-file==1.5.1
rsa==4.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.28
sqlparse==0.2.4
tldextract==3.1.0
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.6
webargs==8.0.0
Werkzeug==2.0.1
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
